+++ src/main/java/com/dp/petshome/persistence/model/User.java
@@ -4,163 +4,162 @@ import java.io.Serializable;
 import java.util.Date;
 
 public class User implements Serializable {
-    private Integer id;
+	private Integer id;
 
-    private String openid;
+	private String openid;
 
-    private Integer role;
+	private Integer role;
 
-    private String name;
+	private String name;
 
-    private String nickname;
+	private String nickname;
 
-    private Integer sex;
+	private Integer sex;
 
-    private String headImg;
+	private String headImg;
 
-    private Integer score;
+	private Integer score;
 
-    private Integer balance;
+	private Double balance;
 
-    private String tel;
+	private String tel;
 
-    private String country;
+	private String country;
 
-    private String province;
+	private String province;
 
-    private String city;
+	private String city;
 
-    private Date createTime;
+	private Date createTime;
 
-    private Date updateTime;
+	private Date updateTime;
 
-    private static final long serialVersionUID = 1L;
+	private static final long serialVersionUID = 1L;
 
-    public Integer getId() {
-        return id;
-    }
+	public Integer getId() {
+		return id;
+	}
 
-    public void setId(Integer id) {
-        this.id = id;
-    }
+	public void setId(Integer id) {
+		this.id = id;
+	}
 
-    public String getOpenid() {
-        return openid;
-    }
+	public String getOpenid() {
+		return openid;
+	}
 
-    public void setOpenid(String openid) {
-        this.openid = openid == null ? null : openid.trim();
-    }
+	public void setOpenid(String openid) {
+		this.openid = openid == null ? null : openid.trim();
+	}
 
-    public Integer getRole() {
-        return role;
-    }
+	public Integer getRole() {
+		return role;
+	}
 
-    public void setRole(Integer role) {
-        this.role = role;
-    }
+	public void setRole(Integer role) {
+		this.role = role;
+	}
 
-    public String getName() {
-        return name;
-    }
+	public String getName() {
+		return name;
+	}
 
-    public void setName(String name) {
-        this.name = name == null ? null : name.trim();
-    }
+	public void setName(String name) {
+		this.name = name == null ? null : name.trim();
+	}
 
-    public String getNickname() {
-        return nickname;
-    }
+	public String getNickname() {
+		return nickname;
+	}
 
-    public void setNickname(String nickname) {
-        this.nickname = nickname == null ? null : nickname.trim();
-    }
+	public void setNickname(String nickname) {
+		this.nickname = nickname == null ? null : nickname.trim();
+	}
 
-    public Integer getSex() {
-        return sex;
-    }
+	public Integer getSex() {
+		return sex;
+	}
 
-    public void setSex(Integer sex) {
-        this.sex = sex;
-    }
+	public void setSex(Integer sex) {
+		this.sex = sex;
+	}
 
-    public String getHeadImg() {
-        return headImg;
-    }
+	public String getHeadImg() {
+		return headImg;
+	}
 
-    public void setHeadImg(String headImg) {
-        this.headImg = headImg == null ? null : headImg.trim();
-    }
+	public void setHeadImg(String headImg) {
+		this.headImg = headImg == null ? null : headImg.trim();
+	}
 
-    public Integer getScore() {
-        return score;
-    }
+	public Integer getScore() {
+		return score;
+	}
 
-    public void setScore(Integer score) {
-        this.score = score;
-    }
+	public void setScore(Integer score) {
+		this.score = score;
+	}
 
-    public Integer getBalance() {
-        return balance;
-    }
+	public Double getBalance() {
+		return balance;
+	}
 
-    public void setBalance(Integer balance) {
-        this.balance = balance;
-    }
+	public void setBalance(Double balance) {
+		this.balance = balance;
+	}
 
-    public String getTel() {
-        return tel;
-    }
+	public String getTel() {
+		return tel;
+	}
 
-    public void setTel(String tel) {
-        this.tel = tel == null ? null : tel.trim();
-    }
+	public void setTel(String tel) {
+		this.tel = tel == null ? null : tel.trim();
+	}
 
-    public String getCountry() {
-        return country;
-    }
+	public String getCountry() {
+		return country;
+	}
 
-    public void setCountry(String country) {
-        this.country = country == null ? null : country.trim();
-    }
+	public void setCountry(String country) {
+		this.country = country == null ? null : country.trim();
+	}
 
-    public String getProvince() {
-        return province;
-    }
+	public String getProvince() {
+		return province;
+	}
 
-    public void setProvince(String province) {
-        this.province = province == null ? null : province.trim();
-    }
+	public void setProvince(String province) {
+		this.province = province == null ? null : province.trim();
+	}
 
-    public String getCity() {
-        return city;
-    }
+	public String getCity() {
+		return city;
+	}
 
-    public void setCity(String city) {
-        this.city = city == null ? null : city.trim();
-    }
+	public void setCity(String city) {
+		this.city = city == null ? null : city.trim();
+	}
 
-    public Date getCreateTime() {
-        return createTime;
-    }
+	public Date getCreateTime() {
+		return createTime;
+	}
 
-    public void setCreateTime(Date createTime) {
-        this.createTime = createTime;
-    }
+	public void setCreateTime(Date createTime) {
+		this.createTime = createTime;
+	}
 
-    public Date getUpdateTime() {
-        return updateTime;
-    }
+	public Date getUpdateTime() {
+		return updateTime;
+	}
 
-    public void setUpdateTime(Date updateTime) {
-        this.updateTime = updateTime;
-    }
+	public void setUpdateTime(Date updateTime) {
+		this.updateTime = updateTime;
+	}
 
 	@Override
 	public String toString() {
-		return "User [id=" + id + ", openid=" + openid + ", role=" + role + ", name=" + name + ", nickname=" + nickname
-				+ ", sex=" + sex + ", headImg=" + headImg + ", score=" + score + ", balance=" + balance + ", tel=" + tel
-				+ ", country=" + country + ", province=" + province + ", city=" + city + ", createTime=" + createTime
+		return "User [id=" + id + ", openid=" + openid + ", role=" + role + ", name=" + name + ", nickname=" + nickname + ", sex=" + sex + ", headImg=" + headImg + ", score="
+				+ score + ", balance=" + balance + ", tel=" + tel + ", country=" + country + ", province=" + province + ", city=" + city + ", createTime=" + createTime
 				+ ", updateTime=" + updateTime + "]";
 	}
 }
